requirements.txt
# FinTech Data Platform Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Dashboard and visualization
dash>=3.0.0
plotly>=5.0.0
dash-bootstrap-components>=2.0.0

# API framework
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Logging and configuration
structlog>=23.0.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# File watching and streaming
watchdog>=3.0.0
kafka-python>=2.0.2